[RENAME] notification list 이름 구분을 위한 패키지 추가 https://github.com/POLZZAK/POLZZAK-Android/issues/68
