--- app/src/main/java/com/polzzak_android/presentation/feature/notification/kid/KidNotificationFragment.kt
@@ -1,7 +1,0 @@
-package com.polzzak_android.presentation.feature.notification.kid
-
-import com.polzzak_android.presentation.feature.notification.base.BaseNotificationFragment
-
-class KidNotificationFragment:BaseNotificationFragment(){
-
-}
