+++ app/src/main/java/com/polzzak_android/presentation/feature/notification/list/NotificationItemStateController.kt
@@ -1,4 +1,4 @@
-package com.polzzak_android.presentation.feature.notification
+package com.polzzak_android.presentation.feature.notification.list
 
 interface NotificationItemStateController {
     fun setHorizontalScrollPosition(id: Int, position: Int)
