+++ app/src/main/java/com/polzzak_android/presentation/feature/notification/list/kid/KidNotificationListFragment.kt
@@ -1,0 +1,7 @@
+package com.polzzak_android.presentation.feature.notification.list.kid
+
+import com.polzzak_android.presentation.feature.notification.list.base.BaseNotificationListFragment
+
+class KidNotificationListFragment: BaseNotificationListFragment(){
+
+}
