+++ app/src/main/java/com/polzzak_android/presentation/feature/notification/list/model/NotificationsModel.kt
@@ -1,4 +1,4 @@
-package com.polzzak_android.presentation.feature.notification.model
+package com.polzzak_android.presentation.feature.notification.list.model
 
 //TODO API 모델 확인 후 수정 필요
 data class NotificationsModel(
