+++ app/src/main/java/com/polzzak_android/presentation/feature/notification/list/protector/ProtectorNotificationListFragment.kt
@@ -1,0 +1,7 @@
+package com.polzzak_android.presentation.feature.notification.list.protector
+
+import com.polzzak_android.presentation.feature.notification.list.base.BaseNotificationListFragment
+
+class ProtectorNotificationListFragment : BaseNotificationListFragment() {
+
+}
