--- app/src/main/java/com/polzzak_android/presentation/feature/notification/protector/ProtectorNotificationFragment.kt
@@ -1,7 +1,0 @@
-package com.polzzak_android.presentation.feature.notification.protector
-
-import com.polzzak_android.presentation.feature.notification.base.BaseNotificationFragment
-
-class ProtectorNotificationFragment : BaseNotificationFragment() {
-
-}
