+++ app/src/main/res/layout/fragment_kid_notification.xml
@@ -4,7 +4,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.feature.notification.kid.KidNotificationFragment">
+        tools:context=".presentation.feature.notification.list.kid.KidNotificationListFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
